Move TableView class into app.py
